fix: Make settings screen modular
