+++ lib/presentation/widgets/screens/settings/sections/schedule_section_skeleton.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+import 'package:dienstplan/core/l10n/app_localizations.dart';
+import 'package:dienstplan/presentation/widgets/screens/settings/settings_section.dart';
+import 'package:dienstplan/presentation/widgets/common/cards/navigation_card_skeleton.dart';
+
+class ScheduleSectionSkeleton extends StatelessWidget {
+  const ScheduleSectionSkeleton({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final AppLocalizations l10n = AppLocalizations.of(context);
+    return SettingsSection(
+      title: l10n.schedule,
+      cards: const [
+        NavigationCardSkeleton(
+          icon: Icons.calendar_today_outlined,
+          title: 'Duty Schedule',
+          showSubtitleSkeleton: true,
+        ),
+        NavigationCardSkeleton(
+          icon: Icons.favorite_outlined,
+          title: 'My Duty Group',
+          showSubtitleSkeleton: true,
+        ),
+        NavigationCardSkeleton(
+          icon: Icons.view_week_outlined,
+          title: 'Calendar Format',
+          showSubtitleSkeleton: true,
+        ),
+      ],
+    );
+  }
+}
